refactor(context): fold getActiveEvents into compaction_utils
`compactor_utils.ts` landed one directory entry away from the existing
`compaction_utils.ts`, which already held the sibling compaction helper
(`calculateRetainStartIndex`). Two modules whose names differ by three
characters is a trap, and `token_based_context_compactor.ts` was importing
from both on adjacent lines. Neither module is re-exported, so this is an
internal move with no API-surface change.

`AgentControlledContextCompactor` still carried a byte-identical private copy
of `getActiveEvents`, including the same redundant
`as CompactedEvent | undefined` cast. It now uses the shared helper, so the
cast is gone at the root rather than in one of two copies.
